modify little bug about front-end rendering

+++ nesplash/static/js/common.js
@@ -11,20 +11,22 @@ class Common {
     // change navbar info
     statusChange(data){
         // signin
-        if(data.role_id == "3"){
-            const a = document.createElement("a");
-            const admin = signin.parentNode.insertBefore(a,signin)
-            admin.classList.add("navbar-brand-2");
-            admin.setAttribute("href", "/admin")
-            admin.textContent = "Admin"
+        if(data.check == true){
+            if(data.role_id == "3"){
+                const a = document.createElement("a");
+                const admin = signin.parentNode.insertBefore(a,signin)
+                admin.classList.add("navbar-brand-2");
+                admin.setAttribute("href", "/admin")
+                admin.textContent = "Admin"
+            }
+            signin.setAttribute("id", "account");
+            signin.setAttribute("href", "/account/data");
+            signin.textContent = `${data.username}`;
+            // signup
+            signup.removeAttribute("href");
+            signup.setAttribute("id", "logout");
+            signup.textContent = "logout";
         }
-        signin.setAttribute("id", "account");
-        signin.setAttribute("href", "/account/data");
-        signin.textContent = `${data.username}`;
-        // signup
-        signup.removeAttribute("href");
-        signup.setAttribute("id", "logout");
-        signup.textContent = "logout";
     }
 
     // check status
@@ -38,23 +40,24 @@ class Common {
             
             if(result.message != null){
                 this.statusChange(result.message)
-                
-                const logout = document.getElementById("logout");
-                logout.onclick = ()=>{
-                    const url = `${window.port}/api/user`
-                    fetch(url, {
-                        method: "DELETE"
-                    })
-                    .then( async (response) => {
-                        return await response.json()
-                    })
-                    .then((result)=>{
-                        if(result.ok == true){
-                            location.href = `${window.port}` + "/" 
-                        }
-                    })
+
+                if (result.message.check == true) {
+                    const logout = document.getElementById("logout");
+                    logout.onclick = ()=>{
+                        const url = `${window.port}/api/user`
+                        fetch(url, {
+                            method: "DELETE"
+                        })
+                        .then( async (response) => {
+                            return await response.json()
+                        })
+                        .then((result)=>{
+                            if(result.ok == true){
+                                location.href = `${window.port}` + "/" 
+                            }
+                        })
+                    }
                 }
-                
             }
             else{
                 // signin
